Translate settings schema
