+++ jupyterlab_server/translations_handler.py
@@ -11,49 +11,11 @@ import traceback
 import tornado
 from tornado import gen
 
-from .server import APIHandler
-from .settings_handler import get_settings
+from .settings_utils import SchemaHandler
 from .translation_utils import get_language_pack, get_language_packs, is_valid_locale, translator
 
 
-SCHEMA_NAME = '@jupyterlab/translation-extension:plugin'
-
-
-def get_current_locale(config):
-    """
-    Get the current locale for given `config`.
-
-    Parameters
-    ----------
-    config: LabConfig
-        The config.
-
-    Returns
-    -------
-    str
-        The current locale string.
-
-    Notes
-    -----
-    If the locale setting is not valid, it will default to "en".
-    """
-    settings, _warnings = get_settings(
-        config.app_settings_dir,
-        config.schemas_dir,
-        config.user_settings_dir,
-        schema_name=SCHEMA_NAME,
-    )
-    current_locale = settings.get("settings", {}).get("locale", "en")
-    if not is_valid_locale(current_locale):
-        current_locale = "en"
-
-    return current_locale
-
-
-class TranslationsHandler(APIHandler):
-
-    def initialize(self, lab_config):
-        self.lab_config = lab_config
+class TranslationsHandler(SchemaHandler):
 
     @gen.coroutine
     @tornado.web.authenticated
@@ -71,7 +33,8 @@ class TranslationsHandler(APIHandler):
         try:
             if locale == "":
                 data, message = get_language_packs(
-                    display_locale=get_current_locale(self.lab_config))
+                    display_locale=self.get_current_locale()
+                )
             else:
                 data, message = get_language_pack(locale)
                 if data == {} and message == "":
